Refactor buildErrors into functional component
