+++ src/client/components/FieldErrors.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { ErrorMap } from "../shared/types";
-import { WriteResource } from "./resource";
+import { ErrorMap } from "../../shared/types";
+import { WriteResource } from "./../resource";
 
 export function getErrorMessage<D, T>(resource: WriteResource<D, T>): string | undefined {
   return "errors" in resource && typeof resource.errors.message === "string"
@@ -14,10 +14,12 @@ export function getFieldErrors<D, T>(resource: WriteResource<D, T>): ErrorMap<D>
     : {};
 }
 
-export function buildErrors<D, F extends keyof D>(
-  field: F,
-  errors: ErrorMap<D>
-): JSX.Element | null {
+interface FieldErrorsProps<D> {
+  readonly field: keyof D;
+  readonly errors: ErrorMap<D>;
+}
+
+export function FieldErrors<D>({ field, errors }: FieldErrorsProps<D>): JSX.Element | null {
   const fieldErrors = errors[field];
   return fieldErrors ? (
     <div style={{ color: "red" }}>
